Use NSInteger instead of int. NSInteger will decide to use int or long depends on system. You usually want to use NSInteger when you don't know what kind of processor architecture your code might run on, so you may for some reason want the largest possible int type, which on 32 bit systems is just an int, while on a 64-bit system it's a long.

+++ KHDataBinding/KHDataBinding.h
@@ -57,7 +57,7 @@
     UIScrollView *refreshScrollView;
     NSAttributedString *refreshTitle1;
     NSAttributedString *refreshTitle2;
-    int refreshState;
+    NSInteger refreshState;
 
     
 }
